updated error handling

+++ server/app/main/service/text_service.py
@@ -18,12 +18,13 @@ def save_new_text(username, data: Dict[str, str]) -> Dict[str, str]:
     # get user_id from provided username 
     user = User.query.filter_by(username=username).first() 
 
-    # # Get user from provided auth token
+    # Get user from provided auth token
     # logged_in_user = Auth.get_logged_in_user(request)[0]['data']
     # print(logged_in_user)
     
     new_text = Text(
         user_id=user.id,
+        # user_id=logged_in_user['user_id'],
         text_id=str(uuid.uuid4()),
         created_on=datetime.datetime.utcnow(),
         text_title=data['text_title'],
@@ -53,18 +54,22 @@ def update_text(username, text_id, data: Dict[str, str]) -> Dict[str, str]:
     user = User.query.filter_by(username=username).first()
     # get row to update 
     row = Text.query.filter_by(text_id=text_id, user_id=user.id).first() 
+    # check if row exists 
+    if bool(row):
+        # update text title and body
+        row.text_title = data['text_title']  # update created_on time too? 
+        row.text_body = data['text_body']
 
-    # update text title and body
-    row.text_title = data['text_title']  # update created_on time too? 
-    row.text_body = data['text_body']
+        db.session.commit() 
 
-    db.session.commit() 
-
-    response_object = {
-        'status': 'success',
-        'message': 'Successfully updated text.'
-    }
-    return response_object
+        response_object = {
+            'status': 'success',
+            'message': 'Successfully updated text.'
+        }
+        return response_object
+    
+    else: 
+        return False 
 
 
 def delete_a_text(username, text_id):
@@ -72,16 +77,21 @@ def delete_a_text(username, text_id):
     user = User.query.filter_by(username=username).first() 
     # get row to delete 
     row = Text.query.filter_by(text_id=text_id, user_id=user.id).first() 
-    
-    db.session.delete(row)
-    db.session.commit()
+    # check if row exists 
+    if bool(row):
+        # delete row
+        db.session.delete(row)
+        db.session.commit()
 
-    response_object = {
-        'status': 'success',
-        'message': 'Successfully deleted text.'
-    }
+        response_object = {
+            'status': 'success',
+            'message': 'Successfully deleted text.'
+        }
+        
+        return response_object
     
-    return response_object
+    else: 
+        return False 
 
 
 def save_changes(data: Text) -> None:
